requirements.txt
flask>=3.1
flask-login>=0.6
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-wtf>=1.2
flask-cors>=4.0
python-dotenv>=1.0
requests>=2.31
Pillow>=10.0
cryptography>=41.0
numpy>=1.26
markdown>=3.5
gunicorn>=21.2
